Added a few enchantments and the Eye of the Guardian artifact, refactored some things and finished(?) the soul bar display

+++ src/main/java/com/infamous/dungeons_gear/items/artifacts/HarvesterItem.java
@@ -53,7 +53,7 @@ public class HarvesterItem extends ArtifactItem implements ISoulGatherer {
     }
 
     @Override
-    public int getActivationCost(ItemStack stack) {
+    public float getActivationCost(ItemStack stack) {
         return 40;
     }
