+++ src/main/java/com/infamous/dungeons_gear/items/artifacts/LightningRodItem.java
@@ -62,7 +62,7 @@ public class LightningRodItem extends ArtifactItem implements ISoulGatherer {
     }
 
     @Override
-    public int getActivationCost(ItemStack stack) {
+    public float getActivationCost(ItemStack stack) {
         return 15;
     }
 }
